+++ apps/code-examples/src/app/code-examples/forms/file-attachment/basic/demo.component.ts
@@ -1,67 +1,76 @@
 import { CommonModule } from '@angular/common';
-import { Component } from '@angular/core';
+import { Component, inject } from '@angular/core';
 import {
+  AbstractControl,
+  FormBuilder,
+  FormControl,
+  FormGroup,
+  FormsModule,
+  ReactiveFormsModule,
+  Validators,
+} from '@angular/forms';
+import {
+  SkyFileAttachmentChange,
+  SkyFileAttachmentClick,
   SkyFileAttachmentModule,
-  SkyFileDropChange,
   SkyFileItem,
-  SkyFileLink,
 } from '@skyux/forms';
-import { SkyStatusIndicatorModule } from '@skyux/indicators';
 
 @Component({
   standalone: true,
   selector: 'app-demo',
   templateUrl: './demo.component.html',
-  imports: [CommonModule, SkyFileAttachmentModule, SkyStatusIndicatorModule],
+  imports: [
+    CommonModule,
+    FormsModule,
+    ReactiveFormsModule,
+    SkyFileAttachmentModule,
+  ],
 })
 export class DemoComponent {
-  protected acceptedTypes = 'image/png,image/jpeg';
-  protected allItems: (SkyFileItem | SkyFileLink)[] = [];
-  protected hintText = '5 MB maximum';
-  protected inlineHelpContent =
-    'Your logo appears in places such as authentication pages, student and parent portals, and extracurricular home pages.';
-  protected labelText = 'Logo image';
-  protected maxFileSize = 5242880;
-  protected rejectedFiles: SkyFileItem[] = [];
-  protected required = true;
-  protected stacked = 'true';
+  protected attachment: FormControl;
+  protected formGroup: FormGroup;
+  protected maxFileSize = 4000000;
+  protected customValidationError: string | undefined;
 
-  #filesToUpload: SkyFileItem[] = [];
-  #linksToUpload: SkyFileLink[] = [];
-
-  protected deleteFile(file: SkyFileItem | SkyFileLink): void {
-    this.#removeFromArray(this.allItems, file);
-    this.#removeFromArray(this.#filesToUpload, file);
-    this.#removeFromArray(this.#linksToUpload, file);
+  get #reactiveFile(): AbstractControl | null {
+    return this.formGroup.get('attachment');
   }
 
-  protected onFilesChanged(change: SkyFileDropChange): void {
-    this.#filesToUpload = this.#filesToUpload.concat(change.files);
-    this.rejectedFiles = change.rejectedFiles;
-    this.allItems = this.allItems.concat(change.files);
+  constructor() {
+    this.attachment = new FormControl(undefined, Validators.required);
+    this.formGroup = inject(FormBuilder).group({
+      attachment: this.attachment,
+    });
   }
 
-  protected onLinkChanged(change: SkyFileLink): void {
-    this.#linksToUpload = this.#linksToUpload.concat(change);
-    this.allItems = this.allItems.concat(change);
-  }
+  protected onFileChange(result: SkyFileAttachmentChange): void {
+    const file = result.file;
 
-  protected validateFile(file: SkyFileItem): string | undefined {
-    return file.file.name.startsWith('a')
-      ? 'Upload a file that does not begin with the letter "a"'
-      : undefined;
-  }
-
-  #removeFromArray(
-    items: (SkyFileItem | SkyFileLink)[],
-    obj: SkyFileItem | SkyFileLink,
-  ): void {
-    if (items) {
-      const index = items.indexOf(obj);
-
-      if (index !== -1) {
-        items.splice(index, 1);
-      }
+    if (file?.errorType) {
+      this.#reactiveFile?.setValue(undefined);
+    } else {
+      this.#reactiveFile?.setValue(file);
     }
+
+    if (file && file.errorType === 'validate') {
+      this.customValidationError = file.errorParam;
+    } else {
+      this.customValidationError = undefined;
+    }
+  }
+
+  protected onFileClick($event: SkyFileAttachmentClick): void {
+    // Ensure we are only attempting to navigate to locally updated data for download.
+    if ($event.file.url.startsWith('data:')) {
+      const link = document.createElement('a');
+      link.download = $event.file.file.name;
+      link.href = $event.file.url;
+      link.click();
+    }
+  }
+
+  protected validateFile(file: SkyFileItem): string {
+    return file.file.name.startsWith('a') ? 'invalidStartingLetter' : '';
   }
 }
