+++ apps/code-examples/src/app/home/home.component.html
@@ -179,9 +179,15 @@
         </ul>
       </li>
       <li>
-        File Attachment
+        File attachment
         <ul>
           <li><a routerLink="forms/file-attachment/basic">Basic</a></li>
+        </ul>
+      </li>
+      <li>
+        File drop
+        <ul>
+          <li><a routerLink="forms/file-drop/basic">Basic</a></li>
         </ul>
       </li>
       <li>
@@ -202,12 +208,6 @@
         <ul>
           <li><a routerLink="forms/selection-box/checkbox">Checkbox</a></li>
           <li><a routerLink="forms/selection-box/radio">Radio</a></li>
-        </ul>
-      </li>
-      <li>
-        Single file attachment
-        <ul>
-          <li><a routerLink="forms/single-file-attachment/basic">Basic</a></li>
         </ul>
       </li>
       <li>
