+++ libs/components/forms/src/lib/modules/file-attachments/file-attachments.module.ts
@@ -4,7 +4,7 @@ import { SkyFileAttachmentModule } from './file-attachment/file-attachment.modul
 import { SkyFileDropModule } from './file-drop/file-drop.module';
 
 /**
- * @deprecated Import either `SkyFileAttachmentModule` or `SkySingleFileAttachmentModule`.
+ * @deprecated Import either `SkyFileAttachmentModule` or `SkyFileDropModule`.
  */
 @NgModule({
   exports: [SkyFileAttachmentModule, SkyFileDropModule],
